Move serialization code under SAGESharp.SLB to SAGESharp.SLB.IO
This is part of a cleanup to add classes for all SLB
files, this IO code is anyway going to go away with
the new "generic serializers"
