+++ SAGESharp/SLB/IO/ISLBBinaryWriter.cs
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-namespace SAGESharp.SLB
+namespace SAGESharp.SLB.IO
 {
     /// <summary>
     /// Interface that represents an object to write SLB objects as binary data.
