+++ SAGESharp/SLB/Level/Conversation/CharacterBinaryWriter.cs
@@ -3,6 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
+using SAGESharp.SLB.IO;
 using System;
 using System.IO;
