+++ SAGESharpTests/SLB/IO/StringBinaryWriterTests.cs
@@ -10,7 +10,7 @@ using SAGESharp.Testing;
 using System;
 using System.IO;
 
-namespace SAGESharp.SLB
+namespace SAGESharp.SLB.IO
 {
     class StringBinaryWriterTests
     {
