+++ SAGESharpTests/SLB/Level/Conversation/ConversationFooterGeneratorTests.cs
@@ -5,6 +5,7 @@
  */
 using FluentAssertions;
 using NUnit.Framework;
+using SAGESharp.SLB.IO;
 using SAGESharp.Testing;
 using System;
 using System.Collections.Generic;
