+++ SAGESharpTests/SLB/Level/Conversation/FrameBinaryWriterTests.cs
@@ -6,6 +6,7 @@
 using FluentAssertions;
 using NSubstitute;
 using NUnit.Framework;
+using SAGESharp.SLB.IO;
 using SAGESharp.Testing;
 using System;
 using System.IO;
